requirements.txt
gradio
transformers
torch
Pillow
soundfile
torchaudio
imageio
imageio-ffmpeg
